Update python file and tests
