--- .github/workflows/update-tool-folders.yml
@@ -1,26 +1,0 @@
-name: Update Tool Folders
-
-on:
-  workflow_dispatch:
-
-jobs:
-  update-folders:
-    runs-on: ubuntu-latest
-
-    steps:
-    - name: Checkout repository
-      uses: actions/checkout@v2
-
-    - name: Set up Python
-      uses: actions/setup-python@v2
-      with:
-        python-version: '3.x'
-
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        # Add any additional dependencies if required
-        # pip install <dependency>
-
-    - name: Run update tool
-      run: python ./update-tool-folders.py
